Explicit check that num_mythic_body_parts != 6
The reason this is necessary is that metadata cache updates
independently of mythic body parts cached. Although the data is
eventually consistent, this leads to transient overestimates of points
on the leaderboard.

This fix makes it so that any unicorn which has 6 mythic parts is
considered to be an egg and is not counted in the leaderboard.